+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2024-04-04]
+## [1.1.0 - 2024-04-05]
 ### Added
 - Dockerfile
 - Single Spa
@@ -54,3 +54,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - EditProjectParam
 - Bridgeheads list for adding new bridgeheads to a project
 - Bridgehead can accept only if not accepted and reject only if not rejected
+- Fetch labels of script, application form and votum